Added more unit tests for handler
