+++ memcrs/src/memcache_server/handler.rs
@@ -168,20 +168,26 @@ impl BinaryHandler {
 
     fn append_prepend(
         &self,
-        append_req: network::AppendRequest,
+        append_prepend_req: network::AppendRequest,
         response_header: &mut network::ResponseHeader,
     ) -> encoder::BinaryResponse {
-        let record = store::Record::new(append_req.value, append_req.header.cas, 0, 0);
-        let result = if self.is_append(append_req.header.opcode) {
-            self.storage.append(append_req.key, record)
+        let record = store::Record::new(append_prepend_req.value, append_prepend_req.header.cas, 0, 0);
+        let is_append = self.is_append(append_prepend_req.header.opcode);
+        let result = if is_append {
+            self.storage.append(append_prepend_req.key, record)
         } else {
-            self.storage.prepend(append_req.key, record)
+            self.storage.prepend(append_prepend_req.key, record)
         };
 
         match result {
             Ok(status) => {
                 response_header.cas = status.cas;
-                encoder::BinaryResponse::Append(network::AppendResponse {
+                if is_append {
+                    return encoder::BinaryResponse::Append(network::AppendResponse {
+                        header: *response_header,
+                    })
+                }
+                encoder::BinaryResponse::Prepend(network::PrependResponse {
                     header: *response_header,
                 })
             }
@@ -339,802 +345,4 @@ impl BinaryHandler {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::decoder;
-    use super::network;
-    use super::*;
-    use crate::cache::error;
-    use crate::mock::handler::*;
-    use crate::mock::value::from_string;
-    use test_case::test_case;
-
-    use bytes::Bytes;
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn get_request_should_return_not_found_when_not_exists(handler: BinaryHandler) {
-        let key = Bytes::from("key");
-        let header = create_header(network::Command::Get, &key);
-
-        let request = decoder::BinaryRequest::Get(network::GetRequest { header, key });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Error(response) = resp {
-                    assert_eq!(response.header.status, error::CacheError::NotFound as u16);
-                    assert_eq!(response.error, "Not found");
-                    assert_eq!(response.header.body_length, response.error.len() as u32);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn get_quiet_request_should_return_none_when_not_exists(handler: BinaryHandler) {
-        let key = Bytes::from("key");
-        let header = create_header(network::Command::GetQuiet, &key);
-
-        let request = decoder::BinaryRequest::GetQuietly(network::GetQuietRequest { header, key });
-
-        let result = handler.handle_request(request);
-        assert!(result.is_none());
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn get_quiet_key_request_should_return_none_when_not_exists(handler: BinaryHandler) {
-        let key = Bytes::from("key");
-        let header = create_header(network::Command::GetQuiet, &key);
-
-        let request =
-            decoder::BinaryRequest::GetKeyQuietly(network::GetKeyQuietRequest { header, key });
-
-        let result = handler.handle_request(request);
-        assert!(result.is_none());
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn get_key_request_should_return_key_and_record(handler: BinaryHandler) {
-        let key = Bytes::from("test_key");
-        let value = from_string("test value");
-
-        insert_value(&handler, key.clone(), value.clone());
-
-        let header = create_header(network::Command::GetKey, &key);
-        let request = decoder::BinaryRequest::GetKey(network::GetKeyRequest {
-            header,
-            key: key.clone(),
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Get(response) = resp {
-                    assert_ne!(response.header.cas, 0);
-                    check_header(
-                        &response.header,
-                        network::Command::GetKey,
-                        key.len() as u16,
-                        EXTRAS_LENGTH,
-                        0,
-                        0,
-                        key.len() as u32 + value.len() as u32 + EXTRAS_LENGTH as u32,
-                    );
-                    assert_eq!(response.key[..], key[..]);
-                    assert_eq!(response.value[..], value[..]);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn get_quiet_key_request_should_return_key_and_record(handler: BinaryHandler) {
-        let key = Bytes::from("test_key");
-        let value = from_string("test value");
-
-        insert_value(&handler, key.clone(), value.clone());
-
-        let header = create_header(network::Command::GetKeyQuiet, &key);
-        let request = decoder::BinaryRequest::GetKeyQuietly(network::GetKeyQuietRequest {
-            header,
-            key: key.clone(),
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Get(response) = resp {
-                    assert_ne!(response.header.cas, 0);
-                    check_header(
-                        &response.header,
-                        network::Command::GetKeyQuiet,
-                        key.len() as u16,
-                        EXTRAS_LENGTH,
-                        0,
-                        0,
-                        key.len() as u32 + value.len() as u32 + EXTRAS_LENGTH as u32,
-                    );
-                    assert_eq!(response.key[..], key[..]);
-                    assert_eq!(response.value[..], value[..]);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn get_request_should_return_record(handler: BinaryHandler) {
-        let key = Bytes::from("key");
-        let header = create_header(network::Command::Get, &key);
-        const FLAGS: u32 = 0xDEAD_BEEF;
-        let value = from_string("value");
-        let record = store::Record::new(value.clone(), 0, FLAGS, 0);
-
-        let set_result = handler.storage.set(key.clone(), record);
-        assert!(set_result.is_ok());
-
-        let request = decoder::BinaryRequest::Get(network::GetRequest { header, key });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Get(response) = resp {
-                    assert_eq!(response.flags, FLAGS);
-                    assert_ne!(response.header.cas, 0);
-                    check_header(
-                        &response.header,
-                        network::Command::Get,
-                        0,
-                        EXTRAS_LENGTH,
-                        0,
-                        0,
-                        value.len() as u32 + EXTRAS_LENGTH as u32,
-                    );
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn set_request_should_succeed(handler: BinaryHandler) {
-        let key = Bytes::from("key");
-        let header = create_header(network::Command::Set, &key);
-        const FLAGS: u32 = 0xDEAD_BEEF;
-        let value = from_string("value");
-        let request = decoder::BinaryRequest::Set(network::SetRequest {
-            header,
-            flags: FLAGS,
-            expiration: 0,
-            key,
-            value,
-        });
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Set(response) = resp {
-                    assert_ne!(response.header.cas, 0);
-                    check_header(&response.header, network::Command::Set, 0, 0, 0, 0, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn set_request_should_return_item_too_large_(handler: BinaryHandler) {
-        let key = Bytes::from("key");
-        let header = create_header(network::Command::Set, &key);
-        const FLAGS: u32 = 0xDEAD_BEEF;
-        let value = from_string("value");
-        let request = decoder::BinaryRequest::ItemTooLarge(network::SetRequest {
-            header,
-            flags: FLAGS,
-            expiration: 0,
-            key,
-            value,
-        });
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Error(response) = resp {
-                    assert_eq!(response.header.cas, 0);
-                    check_header(
-                        &response.header,
-                        network::Command::Set,
-                        0,
-                        0,
-                        0,
-                        error::CacheError::ValueTooLarge as u16,
-                        response.error.len() as u32,
-                    );
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn set_request_on_cas_mismatch_should_return_key_exists(handler: BinaryHandler) {
-        let key = Bytes::from("key");
-        let mut header = create_header(network::Command::Set, &key);
-        const FLAGS: u32 = 0xDEAD_BEEF;
-        let value = from_string("value");
-
-        let request = decoder::BinaryRequest::Set(network::SetRequest {
-            header,
-            flags: FLAGS,
-            expiration: 0,
-            key: key.clone(),
-            value: value.clone(),
-        });
-
-        let result = handler.handle_request(request);
-
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Set(response) = resp {
-                    assert_ne!(response.header.cas, 0);
-                    check_header(&response.header, network::Command::Set, 0, 0, 0, 0, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-        header.cas = 100;
-        let request = decoder::BinaryRequest::Set(network::SetRequest {
-            header,
-            flags: FLAGS,
-            expiration: 0,
-            key,
-            value,
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Error(response) = resp {
-                    assert_eq!(response.header.cas, 0);
-                    check_header(
-                        &response.header,
-                        network::Command::Set,
-                        0,
-                        0,
-                        0,
-                        error::CacheError::KeyExists as u16,
-                        response.error.len() as u32,
-                    );
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn version_request_should_return_version(handler: BinaryHandler) {
-        let key = String::from("").into_bytes();
-        let header = create_header(network::Command::Version, &key);
-        let request = decoder::BinaryRequest::Version(network::VersionRequest { header });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Version(response) = resp {
-                    check_header(
-                        &response.header,
-                        network::Command::Version,
-                        0,
-                        0,
-                        0,
-                        0,
-                        MEMCRS_VERSION.len() as u32,
-                    );
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn increment_request_should_return_cas(handler: BinaryHandler) {
-        const EXPECTED_VALUE: u64 = 1;
-        let key = Bytes::from("counter");
-        let header = create_header(network::Command::Increment, &key);
-        let request = decoder::BinaryRequest::Increment(network::IncrementRequest {
-            header,
-            delta: 1,
-            initial: 1,
-            expiration: 1,
-            key,
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Increment(response) = resp {
-                    check_header(
-                        &response.header,
-                        network::Command::Increment,
-                        0,
-                        0,
-                        0,
-                        0,
-                        std::mem::size_of::<store::DeltaResultValueType>() as u32,
-                    );
-                    assert_eq!(response.value, EXPECTED_VALUE);
-                    assert_ne!(response.header.cas, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn increment_request_should_increment_value(handler: BinaryHandler) {
-        const EXPECTED_VALUE: u64 = 101;
-        let key = Bytes::from("counter");
-        let value = from_string("100");
-        insert_value(&handler, key.clone(), value);
-
-        let header = create_header(network::Command::Increment, &key);
-        let request = decoder::BinaryRequest::Increment(network::IncrementRequest {
-            header,
-            delta: 1,
-            initial: 1,
-            expiration: 1,
-            key,
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Increment(response) = resp {
-                    check_header(
-                        &response.header,
-                        network::Command::Increment,
-                        0,
-                        0,
-                        0,
-                        0,
-                        std::mem::size_of::<store::DeltaResultValueType>() as u32,
-                    );
-                    assert_eq!(response.value, EXPECTED_VALUE);
-                    assert_ne!(response.header.cas, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn increment_quiet_should_increment_value(handler: BinaryHandler) {
-        let key = Bytes::from("counter");
-        let value = from_string("100");
-        insert_value(&handler, key.clone(), value);
-
-        let header = create_header(network::Command::IncrementQuiet, &key);
-        let request = decoder::BinaryRequest::IncrementQuiet(network::IncrementRequest {
-            header,
-            delta: 1,
-            initial: 1,
-            expiration: 1,
-            key: key.clone(),
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(_resp) => unreachable!(),
-            None => {}
-        }
-        let incremented_value = get_value(&handler, key.clone());
-        let expected_value = from_string("101");
-        assert_eq!(incremented_value[..], expected_value[..]);
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn decrement_request_should_return_cas(handler: BinaryHandler) {
-        const EXPECTED_VALUE: u64 = 1;
-        let key = Bytes::from("counter");
-        let header = create_header(network::Command::Decrement, &key);
-        let request = decoder::BinaryRequest::Decrement(network::DecrementRequest {
-            header,
-            delta: 1,
-            initial: 1,
-            expiration: 1,
-            key,
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Decrement(response) = resp {
-                    check_header(
-                        &response.header,
-                        network::Command::Decrement,
-                        0,
-                        0,
-                        0,
-                        0,
-                        std::mem::size_of::<store::DeltaResultValueType>() as u32,
-                    );
-                    assert_eq!(response.value, EXPECTED_VALUE);
-                    assert_ne!(response.header.cas, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn decrement_request_should_decrement_value(handler: BinaryHandler) {
-        const EXPECTED_VALUE: u64 = 99;
-        let key = Bytes::from("counter");
-        let value = from_string("100");
-        insert_value(&handler, key.clone(), value);
-
-        let header = create_header(network::Command::Decrement, &key);
-        let request = decoder::BinaryRequest::Decrement(network::DecrementRequest {
-            header,
-            delta: 1,
-            initial: 1,
-            expiration: 1,
-            key,
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Decrement(response) = resp {
-                    check_header(
-                        &response.header,
-                        network::Command::Decrement,
-                        0,
-                        0,
-                        0,
-                        0,
-                        std::mem::size_of::<store::DeltaResultValueType>() as u32,
-                    );
-                    assert_eq!(response.value, EXPECTED_VALUE);
-                    assert_ne!(response.header.cas, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn decrement_quiet_should_increment_value(handler: BinaryHandler) {
-        let key = Bytes::from("counter");
-        let value = from_string("100");
-        insert_value(&handler, key.clone(), value);
-
-        let header = create_header(network::Command::DecrementQuiet, &key);
-        let request = decoder::BinaryRequest::DecrementQuiet(network::DecrementRequest {
-            header,
-            delta: 1,
-            initial: 1,
-            expiration: 1,
-            key: key.clone(),
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(_resp) => unreachable!(),
-            None => {}
-        }
-        let dec_value = get_value(&handler, key.clone());
-        let expected_value = from_string("99");
-        assert_eq!(dec_value[..], expected_value[..]);
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn increment_request_should_error_when_expiration_is_ffffffff(handler: BinaryHandler) {
-        let key = Bytes::from("counter");
-        let header = create_header(network::Command::Increment, &key);
-        let request = decoder::BinaryRequest::Increment(network::IncrementRequest {
-            header,
-            delta: 1,
-            initial: 1,
-            expiration: 0xffffffff,
-            key,
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Error(response) = resp {
-                    check_header(
-                        &response.header,
-                        network::Command::Increment,
-                        0,
-                        0,
-                        0,
-                        network::ResponseStatus::KeyNotExists as u16,
-                        response.error.len() as u32,
-                    );
-                    assert_eq!(response.header.cas, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn decrement_request_should_error_when_expiration_is_ffffffff(handler: BinaryHandler) {
-        let key = Bytes::from("counter");
-        let header = create_header(network::Command::Decrement, &key);
-        let request = decoder::BinaryRequest::Decrement(network::DecrementRequest {
-            header,
-            delta: 1,
-            initial: 1,
-            expiration: 0xffffffff,
-            key,
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Error(response) = resp {
-                    check_header(
-                        &response.header,
-                        network::Command::Decrement,
-                        0,
-                        0,
-                        0,
-                        network::ResponseStatus::KeyNotExists as u16,
-                        response.error.len() as u32,
-                    );
-                    assert_eq!(response.header.cas, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn flush_should_remove_all(handler: BinaryHandler) {
-        let value = from_string("test value");
-        for key_suffix in 0..100 {
-            let key = Bytes::from(String::from("test_key") + &key_suffix.to_string());
-            insert_value(&handler, key.clone(), value.clone());
-        }
-
-        let key = String::from("").into_bytes();
-        let header = create_header(network::Command::Flush, &key);
-        let request = decoder::BinaryRequest::Flush(network::FlushRequest {
-            header,
-            expiration: 0,
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Flush(response) = resp {
-                    check_header(&response.header, network::Command::Flush, 0, 0, 0, 0, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn delete_should_remove_from_store(handler: BinaryHandler) {
-        let value = from_string("test value");
-        let key = Bytes::from("test_key");
-        insert_value(&handler, key.clone(), value.clone());
-
-        let header = create_header(network::Command::Delete, &key);
-        let request = decoder::BinaryRequest::Delete(network::DeleteRequest {
-            header,
-            key: key.clone(),
-        });
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Delete(response) = resp {
-                    check_header(&response.header, network::Command::Delete, 0, 0, 0, 0, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn delete_should_return_error_if_not_exists(handler: BinaryHandler) {
-        let key = Bytes::from("test_key");
-
-        let header = create_header(network::Command::DeleteQuiet, &key);
-        let request = decoder::BinaryRequest::DeleteQuiet(network::DeleteRequest {
-            header,
-            key: key.clone(),
-        });
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Error(response) = resp {
-                    check_header(
-                        &response.header,
-                        network::Command::DeleteQuiet,
-                        0,
-                        0,
-                        0,
-                        network::ResponseStatus::KeyNotExists as u16,
-                        response.error.len() as u32,
-                    );
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn noop_request(handler: BinaryHandler) {
-        let key = String::from("").into_bytes();
-
-        let header = create_header(network::Command::Noop, &key);
-        let request = decoder::BinaryRequest::Noop(network::NoopRequest { header });
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Noop(response) = resp {
-                    check_header(&response.header, network::Command::Noop, 0, 0, 0, 0, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn quit_request(handler: BinaryHandler) {
-        let key = String::from("").into_bytes();
-
-        let header = create_header(network::Command::Quit, &key);
-        let request = decoder::BinaryRequest::Quit(network::QuitRequest { header });
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Quit(response) = resp {
-                    check_header(&response.header, network::Command::Quit, 0, 0, 0, 0, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn quit_quiet_request(handler: BinaryHandler) {
-        let key = String::from("").into_bytes();
-
-        let header = create_header(network::Command::QuitQuiet, &key);
-        let request = decoder::BinaryRequest::QuitQuietly(network::QuitRequest { header });
-        let result = handler.handle_request(request);
-        match result {
-            Some(_resp) => unreachable!(),
-            None => {}
-        }
-    }
-
-    #[test_case(create_moka_handler() ; "moka_backend")]
-    #[test_case(create_dash_map_handler() ; "dash_map_backend")]
-    fn add_request(handler: BinaryHandler) {
-        let key = Bytes::from("key");
-        let mut header = create_header(network::Command::Add, &key);
-        const FLAGS: u32 = 0xDEAD_BEEF;
-        let value = from_string("value");
-
-        let request = decoder::BinaryRequest::Add(network::AddRequest {
-            header,
-            flags: FLAGS,
-            expiration: 0,
-            key: key.clone(),
-            value: value.clone(),
-        });
-
-        let result = handler.handle_request(request);
-
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Set(response) = resp {
-                    assert_ne!(response.header.cas, 0);
-                    check_header(&response.header, network::Command::Add, 0, 0, 0, 0, 0);
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-        header.cas = 100;
-        let request = decoder::BinaryRequest::Add(network::SetRequest {
-            header,
-            flags: FLAGS,
-            expiration: 0,
-            key,
-            value,
-        });
-
-        let result = handler.handle_request(request);
-        match result {
-            Some(resp) => {
-                if let encoder::BinaryResponse::Error(response) = resp {
-                    assert_eq!(response.header.cas, 0);
-                    check_header(
-                        &response.header,
-                        network::Command::Add,
-                        0,
-                        0,
-                        0,
-                        error::CacheError::KeyExists as u16,
-                        response.error.len() as u32,
-                    );
-                } else {
-                    unreachable!();
-                }
-            }
-            None => unreachable!(),
-        }
-    }
-}
+mod handler_tests;
